review
- more structured `but-core` addition
- idempotent `add_conflict_markers`
- simplify marker removal
