+++ crates/but-rebase/src/graph_rebase/cherry_pick.rs
@@ -305,21 +305,16 @@ fn commit_from_unconflicted_tree<'repo>(
     let repo = to_rebase.id.repo;
 
     let headers = to_rebase.headers();
-    let was_conflicted = to_rebase.is_conflicted();
     let mut new_commit = to_rebase.inner;
     new_commit.tree = resolved_tree_id.detach();
 
     // Ensure the commit isn't thinking it's conflicted.
-    if was_conflicted {
-        // Strip the legacy header if present
-        if let Some(pos) = new_commit
-            .extra_headers()
-            .find_pos(HEADERS_CONFLICTED_FIELD)
-        {
-            new_commit.extra_headers.remove(pos);
-        }
-        // Strip conflict markers from the commit message
-        new_commit.message = but_core::commit::strip_conflict_markers(new_commit.message.as_ref());
+    new_commit.message = but_core::commit::strip_conflict_markers(new_commit.message.as_ref());
+    if let Some(pos) = new_commit
+        .extra_headers()
+        .find_pos(HEADERS_CONFLICTED_FIELD)
+    {
+        new_commit.extra_headers.remove(pos);
     } else if headers.is_none() {
         let headers = Headers::from_config(&repo.config_snapshot());
         new_commit
@@ -390,10 +385,8 @@ fn commit_from_conflicted_tree<'repo>(
     to_rebase.parents = parents.into();
 
     // Add conflict markers to the commit message
-    if !but_core::commit::message_is_conflicted(to_rebase.inner.message.as_ref()) {
-        to_rebase.inner.message =
-            but_core::commit::add_conflict_markers(to_rebase.inner.message.as_ref());
-    }
+    to_rebase.inner.message =
+        but_core::commit::add_conflict_markers(to_rebase.inner.message.as_ref());
 
     to_rebase.set_headers(&headers);
     Ok(crate::commit::create(
