+++ crates/gitbutler-commit/src/commit_ext.rs
@@ -1,5 +1,4 @@
 use bstr::BStr;
-use but_core::commit::message_is_conflicted;
 use but_core::{ChangeId, commit::Headers};
 
 /// Extension trait for `gix::Commit`.
@@ -28,16 +27,11 @@ impl CommitExt for gix::Commit<'_> {
     }
 
     fn is_conflicted(&self) -> bool {
-        // Check commit message first (new style), fall back to header (legacy).
-        if let Ok(commit) = self.decode() {
-            if message_is_conflicted(commit.message) {
-                return true;
-            }
-            Headers::try_from_commit_headers(|| commit.extra_headers())
-                .is_some_and(|hdr| hdr.is_conflicted())
-        } else {
-            false
-        }
+        let Ok(commit) = self.decode() else {
+            return false;
+        };
+        let headers = Headers::try_from_commit_headers(|| commit.extra_headers());
+        but_core::commit::is_conflicted(commit.message, headers.as_ref())
     }
 }
